Add usePageTitles hook
